--- a/sessionmanager/sessionmanager.go
+++ b/sessionmanager/sessionmanager.go
@@ -23,10 +23,14 @@ type Session interface {
 type sesTrk struct {
 	session Session
 	pm      bssession.PeerManager
+	srs     bssession.RequestSplitter
 }
 
 // SessionFactory generates a new session for the SessionManager to track.
-type SessionFactory func(ctx context.Context, id uint64, pm bssession.PeerManager) Session
+type SessionFactory func(ctx context.Context, id uint64, pm bssession.PeerManager, srs bssession.RequestSplitter) Session
+
+// RequestSplitterFactory generates a new request splitter for a session.
+type RequestSplitterFactory func(ctx context.Context) bssession.RequestSplitter
 
 // PeerManagerFactory generates a new peer manager for a session.
 type PeerManagerFactory func(ctx context.Context, id uint64) bssession.PeerManager
@@ -34,9 +38,11 @@ type PeerManagerFactory func(ctx context.Context, id uint64) bssession.PeerManag
 // SessionManager is responsible for creating, managing, and dispatching to
 // sessions.
 type SessionManager struct {
-	ctx                context.Context
-	sessionFactory     SessionFactory
-	peerManagerFactory PeerManagerFactory
+	ctx                    context.Context
+	sessionFactory         SessionFactory
+	peerManagerFactory     PeerManagerFactory
+	requestSplitterFactory RequestSplitterFactory
+
 	// Sessions
 	sessLk   sync.Mutex
 	sessions []sesTrk
@@ -47,11 +53,12 @@ type SessionManager struct {
 }
 
 // New creates a new SessionManager.
-func New(ctx context.Context, sessionFactory SessionFactory, peerManagerFactory PeerManagerFactory) *SessionManager {
+func New(ctx context.Context, sessionFactory SessionFactory, peerManagerFactory PeerManagerFactory, requestSplitterFactory RequestSplitterFactory) *SessionManager {
 	return &SessionManager{
-		ctx:                ctx,
-		sessionFactory:     sessionFactory,
-		peerManagerFactory: peerManagerFactory,
+		ctx:                    ctx,
+		sessionFactory:         sessionFactory,
+		peerManagerFactory:     peerManagerFactory,
+		requestSplitterFactory: requestSplitterFactory,
 	}
 }
 
@@ -62,8 +69,9 @@ func (sm *SessionManager) NewSession(ctx context.Context) exchange.Fetcher {
 	sessionctx, cancel := context.WithCancel(ctx)
 
 	pm := sm.peerManagerFactory(sessionctx, id)
-	session := sm.sessionFactory(sessionctx, id, pm)
-	tracked := sesTrk{session, pm}
+	srs := sm.requestSplitterFactory(sessionctx)
+	session := sm.sessionFactory(sessionctx, id, pm, srs)
+	tracked := sesTrk{session, pm, srs}
 	sm.sessLk.Lock()
 	sm.sessions = append(sm.sessions, tracked)
 	sm.sessLk.Unlock()