… sessionmanager/sessionmanager_test.go
… sessionpeermanager/sessionpeermanager_test.go
… sessionrequestsplitter/sessionrequestsplitter.go
… sessionrequestsplitter/sessionrequestsplitter_test.go