# Goals

This is an initial PR to speed up sessions and reduce dupes. It is not the ONLY improvement I'm working on, but it makes significant progress. It matches Jeremy's PR on benchmarks for dupes (for the most part), but matches master on speed (for the most part). In addition, it shows impressive results with `LS` on Wikipedia. See results in comments.

# Implementation

This PR has basically 3 improvements:

1. The session peer manager is just SLIGHTLY smarter in that:
   a. It returns a maximum of 32 peers (why would you ever need to ask more at once?)
   b. It somewhat prioritizes the best peers. The strategy is simply if you receive a block from a peer first, you assume that's a good peer and put it at the front of the list for peers on your next want request. This is not ordering peers by latency, it's just assuming whoever responds is a good person to ask next.

2. The split broadcast vs target max want list for sessions from jeremy's code is copied over. Basically, this is just the assumption that you broadcast only 4 wants at once (previously 16), but once you have peers, you go up to 32 on targeted requests

3. Requests for a series of blocks are split among peers, similarly to Jeremy's PR. However, where Jeremy's PR kept the splitting right in the Session, this moves the splitting out to a separate SessionRequestSplitter module. The strategy pursued is slightly different, and is simply attempting to minimize dupes w/o slowing down the overall request speed by:
   a. If dupes go above a certain threshold, increase the split factor
   b. If they fall below a certain threshold, decrease the split factor (to increase speed, just in case)
Honestly, there are a lot of better ways to do this, but this is functional and works well. (though it vacillates between extremes, cause there's minimal throttling of these adjustments)
   c. To make this work, dupes are now tracked on a per session basis (and we keep a list of past wants to know what's a dupe)

# Next

Once this PR is merged, there are three further optimizations:
1. Peer optimization should be tracked more closely by monitoring actual latency on requests
2. Request splitting should be more intelligent about reducing dupes, and throttled on its adjustments, and possibly become more "fine-grained" (as in fewer adjustments, or smaller adjustments) as a session is open longer, cause at that point, you know more about what the network looks like for whatever it is you're requesting.
3. Request splitting should also take into account how many requests a given node has at once -- that way we don't overload the pipe on a given peer. This first needs: https://github.com/ipfs/go-bitswap/issues/40

# For discussion

This is not a "end goal" PR for session speedup, but a first step and a framework for further optimizations (basically, improvements to the peer manager and the request splitter can proceed independently). I actually have progress on some of the above next steps, but I feel like that should be in a separate PR, so the number of changes aren't overwhelming, and also, I want to make sure this approach feels ok to people.

This also includes some minor changes to the benchmarks, primarily to put the saved json file in a folder that isn't included in git commits.