@hannahhoward 2018-12-04
Progress tracking:

 Name                                                                                          | Dups-Master | Dups-Jeremy | Dups-New | MsgSent-master | MsgSent-Jeremy | MsgSent-New | MsgRecd-Master | MsgRecd-Jeremy | MsgRecd-New | Time-Master | Time-Jeremy | Time-new | 
|-----------------------------------------------------------------------------------------------|-------------|-------------|----------|----------------|----------------|-------------|----------------|----------------|-------------|-------------|-------------|----------| 
| BenchmarkDups2Nodes/AllToAll-OneAtATime                                                       | 1           | 1           | 1        | 199            | 200            | 198         | 101            | 101            | 101         | 2s 202ms    | 2s 209ms    | 2s 248ms | 
| BenchmarkDups2Nodes/AllToAll-BigBatch                                                         | 16          | 4           | 16       | 85             | 138            | 145         | 60             | 85             | 77          | 159ms       | 548ms       | 163ms    | 
| BenchmarkDups2Nodes/Overlap1-OneAtATime                                                       | 0           | 0           | 0        | 251            | 249            | 251         | 100            | 100            | 100         | 2s 766ms    | 2s 788ms    | 2s 835ms | 
| BenchmarkDups2Nodes/Overlap2-BatchBy10                                                        | 34          | 32          | 34       | 131            | 132            | 189         | 20             | 24             | 20          | 221ms       | 857ms       | 221ms    | 
| BenchmarkDups2Nodes/Overlap3-OneAtATime                                                       | 34          | 32          | 34       | 394            | 392            | 390         | 134            | 132            | 134         | 2s 764ms    | 4s 474ms    | 2s 770ms | 
| BenchmarkDups2Nodes/Overlap3-BatchBy10                                                        | 34          | 32          | 34       | 129            | 129            | 185         | 20             | 25             | 20          | 218ms       | 857ms       | 222ms    | 
| BenchmarkDups2Nodes/Overlap3-AllConcurrent                                                    | 34          | 29          | 29       | 194            | 203            | 241         | 99             | 111            | 104         | 150ms       | 1s 179ms    | 756ms    | 
| BenchmarkDups2Nodes/Overlap3-BigBatch                                                         | 31          | 2           | 31       | 179            | 165            | 269         | 95             | 80             | 104         | 767ms       | 4s 945ms    | 760ms    | 
| BenchmarkDups2Nodes/Overlap3-UnixfsFetch                                                      | 34          | 1           | 31       | 173            | 105            | 235         | 90             | 40             | 98          | 178ms       | 1s 312ms    | 761ms    | 
| BenchmarkDups2Nodes/10Nodes-AllToAll-OneAtATime                                               | 8           | 8           | 8        | 215            | 216            | 214         | 108            | 108            | 108         | 2s 194ms    | 2s 216ms    | 2s 223ms | 
| BenchmarkDups2Nodes/10Nodes-AllToAll-BatchFetchBy10                                           | 80          | 32          | 80       | 89             | 74             | 159         | 18             | 22             | 18          | 223ms       | 268ms       | 227ms    | 
| BenchmarkDups2Nodes/10Nodes-AllToAll-BigBatch                                                 | 128         | 32          | 128      | 141            | 103            | 217         | 70             | 71             | 78          | 155ms       | 554ms       | 163ms    | 
| BenchmarkDups2Nodes/10Nodes-AllToAll-AllConcurrent                                            | 128         | 32          | 128      | 182            | 163            | 177         | 98             | 99             | 78          | 154ms       | 545ms       | 156ms    | 
| BenchmarkDups2Nodes/10Nodes-AllToAll-UnixfsFetch                                              | 8           | 8           | 8        | 82             | 114            | 149         | 55             | 57             | 78          | 176ms       | 115ms       | 180ms    | 
| BenchmarkDups2Nodes/10Nodes-OnePeerPerBlock-OneAtATime                                        | 0           | 0           | 0        | 1604           | 1652           | 1684        | 100            | 100            | 100         | 6s 697ms    | 9s 540ms    | 6s 777ms | 
| BenchmarkDups2Nodes/10Nodes-OnePeerPerBlock-BigBatch                                          | 0           | 0           | 0        | 607            | 885            | 1018        | 80             | 89             | 80          | 1s 55ms     | 12s 626ms   | 1s 65ms  | 
| BenchmarkDups2Nodes/10Nodes-OnePeerPerBlock-UnixfsFetch                                       | 0           | 0           | 0        | 649            | 670            | 834         | 71             | 41             | 74          | 1s 489ms    | 2s 478ms    | 1s 464ms | 
| BenchmarkDups2Nodes/200Nodes-AllToAll-BigBatch                                                | 3168        | 792         | 3152     | 785            | 1026           | 1240        | 201            | 214            | 199         | 384ms       | 322ms       | 429ms    | 
| BenchmarkDupsManyNodesRealWorldNetwork/200Nodes-AllToAll-BigBatch-FastNetwork                 | 5250        | 1192        | 4768     | 3226           | 1871           | 4774        | 863            | 493            | 479         | 1s 705ms    | 4s 602ms    | 1s 927ms | 
| BenchmarkDupsManyNodesRealWorldNetwork/200Nodes-AllToAll-BigBatch-AverageVariableSpeedNetwork | 2560        | 1192        | 2432     | 2995           | 1863           | 4285        | 305            | 493            | 304         | 1s 382ms    | 3s 855ms    | 1s 375ms | 
| BenchmarkDupsManyNodesRealWorldNetwork/200Nodes-AllToAll-BigBatch-SlowVariableSpeedNetwork    | 3240        | 1040        | 2448     | 6127           | 1872           | 2146        | 943            | 454            | 308         | 2s 71ms     | 5s 340ms    | 1s 453ms |