@hannahhoward 2018-12-05
Some pretty exciting results with ls-ing Wikipedia

Master `ipfs bitswap stat` results at end:

```
bitswap status
	provides buffer: 0 / 256
	blocks received: 255280
	blocks sent: 5213
	data received: 142597822
	data sent: 8654147
	dup blocks received: 181268
	dup data received: 99 MB
	wantlist [0 keys]
	partners [637]
```

Master `time` command results:

```
cmd/ipfs/ipfs ls --resolve-type=false   7.45s user 3.51s system 0% cpu 2:14:45.69 total
```

This PR `ipfs bitswap stat` results at end:

```
bitswap status
	provides buffer: 0 / 256
	blocks received: 91229
	blocks sent: 1103
	data received: 53514594
	data sent: 3224259
	dup blocks received: 16839
	dup data received: 10 MB
	wantlist [0 keys]
```

This PR `time` command results:

```
cmd/ipfs/ipfs ls --resolve-type=false   8.27s user 3.91s system 0% cpu 31:02.45 total
```