@hannahhoward 2018-12-05
Table updated:

- Mostly tracking Jeremy PR in terms of dupes (except `BenchmarkDups2Nodes/Overlap3-BigBatch` & `BenchmarkDups2Nodes/Overlap3-UnixfsFetch`)

- Mostly tracking master in terms of time (except Real-world benchmarks, `BenchmarkDups2Nodes/AllToAll-BigBatch`, `BenchmarkDups2Nodes/Overlap2-BatchBy10`,
`BenchmarkDups2Nodes/10Nodes-AllToAll-BigBatch`, `BenchmarkDups2Nodes/10Nodes-AllToAll-AllConcurrent`)